Move messageChannelPolyfill to internal test to prevent from being run as test
